refactor: move persistence classes
